requirements.txt
crewai
crewai-tools
fastapi
uvicorn
python-dotenv
pyttsx3
litellm
huggingface-hub
